Recent fixes to oc_config_validate (#308)
* Update in testcase optional arguments.

Better name for testcase argument that checks delay. This makes it clear what the argument is for.

Also, only checking optional arguments when present.

* Add option to explicitly trust Target TLS Cert.

This exactly fetches the TLS Cert from the Target and uses it as the 
Root CA Cert for the client. 

This effectively trusts Self-Signed TLS Certs from the Client.

* Import directly the Exceptions from grpc library.

This avoids long lines when catching the known gRCP exceptions.

* Updated minor version.